+++ src/helpers/formatting.helpers.ts
@@ -1,5 +1,5 @@
 import getSymbolFromCurrency from 'currency-symbol-map'
-import { BN, BnLike } from '@/utils/math.util'
+import { BN, BnLike } from '@distributedlab/tools'
 
 export function cropAddress(address: string, symbol = 6) {
   return `${address.slice(0, symbol)}...${address.slice(-symbol)}`
@@ -12,16 +12,8 @@ function _formatMoney(formattedAmount: string, currency = '') {
     : `${formattedAmount} ${currency}`
 }
 
-export function formatFiatAsset(amount: BnLike, currency = '') {
-  const formattedAmount = new BN(amount).format({
-    decimals: 2,
-  })
-
-  return _formatMoney(formattedAmount, currency)
-}
-
 export function formatFiatAssetFromWei(amount: BnLike, currency = '') {
-  const formattedAmount = new BN(amount).fromWei().format({
+  const formattedAmount = BN.fromBigInt(amount).format({
     decimals: 2,
   })
 
@@ -33,17 +25,9 @@ export function formatAssetFromWei(
   decimals: number,
   currency = '',
 ) {
-  const formattedAmount = new BN(amount).fromWei().format({ decimals })
+  const formattedAmount = BN.fromBigInt(amount).format({ decimals })
 
   return currency ? `${formattedAmount} ${currency}` : formattedAmount
-}
-
-export function formatNumber(amount: BnLike) {
-  const formattedAmount = new BN(amount).format({
-    decimals: 2,
-  })
-
-  return formattedAmount
 }
 
 export function formatBytes(bytes: number, decimals = 2): string {